[AIP-21] Move contrib.hooks.fs_hook hooks.filesystem
