+++ scripts/ci/pylint_todo.txt
@@ -12,7 +12,7 @@
 ./airflow/contrib/hooks/dingding_hook.py
 ./airflow/contrib/hooks/discord_webhook_hook.py
 ./airflow/contrib/hooks/emr_hook.py
-./airflow/contrib/hooks/fs_hook.py
+./airflow/hooks/filesystem.py
 ./airflow/contrib/hooks/ftp_hook.py
 ./airflow/contrib/hooks/jenkins_hook.py
 ./airflow/contrib/hooks/openfaas_hook.py
